+++ PatchDay/Hormones/HormonesTable.swift
@@ -23,7 +23,7 @@ class HormonesTable: TableViewWrapper<HormoneCell> {
         guard let sdk = sdk else { return }
         for row in 0..<SupportedHormoneUpperQuantityLimit {
             if let cell = dequeueCell() {
-                cell.configure(at: row, sdk)
+                cell.configure(viewModel: HormoneCellViewModel(cellIndex: row, sdk: sdk))
                 _cells.append(cell)
             }
         }