+++ PatchDay/Pills/PillsViewModel.swift
@@ -75,7 +75,7 @@ class PillsViewModel: CodeBehindDependencies<PillsViewModel> {
         guard let pill = sdk?.pills[index] else { return }
         let goToDetails = { self.goToPillDetails(pillIndex: index, pillsViewController: viewController) }
         let takePill = {
-            _ = self.takePill(at: index)
+            self.takePill(at: index)
             takePillCompletion()
             self.tabs?.reflectPills()
             self.badge?.reflect()